refactor: centralize progress and toast components
